Invert dependency between untypedTree and editableTree
