+++ experimental/dds/tree2/src/feature-libraries/editable-tree/index.ts
@@ -4,18 +4,12 @@
  */
 
 export {
-	typeSymbol,
 	EditableTree,
 	EditableField,
 	EditableTreeOrPrimitive,
 	proxyTargetSymbol,
 	UnwrappedEditableTree,
 	UnwrappedEditableField,
-	getField,
-	parentField,
-	EditableTreeEvents,
-	on,
-	contextSymbol,
 	NewFieldContent,
 	areCursors,
 	localNodeKeySymbol,
