+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -9,18 +9,12 @@ export {
 	EditableTreeContext,
 	EditableTreeOrPrimitive,
 	getEditableTreeContext,
-	typeSymbol,
 	isEditableField,
 	isPrimitive,
 	isEditableTree,
 	proxyTargetSymbol,
 	UnwrappedEditableField,
 	UnwrappedEditableTree,
-	getField,
-	parentField,
-	EditableTreeEvents,
-	on,
-	contextSymbol,
 	NewFieldContent,
 	localNodeKeySymbol,
 	createDataBinderBuffering,
@@ -203,6 +197,12 @@ export {
 	UnwrappedUntypedField,
 	UnwrappedUntypedTree,
 	UntypedTreeOrPrimitive,
+	typeSymbol,
+	getField,
+	parentField,
+	EditableTreeEvents,
+	on,
+	contextSymbol,
 } from "./untypedTree";
 
 // Split into separate import and export for compatibility with API-Extractor.
