+++ experimental/dds/tree2/src/feature-libraries/untypedTree.ts
@@ -11,20 +11,11 @@ import {
 	ForestEvents,
 	NamedTreeSchema,
 	SchemaData,
+	UpPath,
+	PathVisitor,
 } from "../core";
 import { ISubscribable } from "../events";
-import { requireAssignableTo } from "../util";
 import { PrimitiveValue, MarkedArrayLike, typeNameSymbol, valueSymbol } from "./contextuallyTyped";
-import {
-	EditableField,
-	EditableTree,
-	EditableTreeEvents,
-	contextSymbol,
-	getField,
-	on,
-	parentField,
-	typeSymbol,
-} from "./editable-tree";
 
 /**
  * This file provides an API for working with trees which is type safe even when schema is not known.
@@ -37,6 +28,39 @@ import {
  * SchemaAware builds on this, adding editing and type safe APIs which can be accessed via SchematizeView.
  * Once this is finished, the unsafe EditableTree types can be removed (or converted to package internal documentation for the proxies).
  */
+
+/**
+ * A symbol to get the type of {@link EditableTree} in contexts where string keys are already in use for fields.
+ * @alpha
+ */
+export const typeSymbol: unique symbol = Symbol("editable-tree:type");
+
+/**
+ * A symbol to get the function, which returns the field of {@link EditableTree} without unwrapping,
+ * in contexts where string keys are already in use for fields.
+ * @alpha
+ */
+export const getField: unique symbol = Symbol("editable-tree:getField()");
+
+/**
+ * A symbol to get information about where an {@link EditableTree} is parented
+ * in contexts where string keys are already in use for fields.
+ * @alpha
+ */
+export const parentField: unique symbol = Symbol("editable-tree:parentField()");
+
+/**
+ * A symbol to get a common context of a "forest" of EditableTrees
+ * in contexts where string keys are already in use for fields.
+ * @alpha
+ */
+export const contextSymbol: unique symbol = Symbol("editable-tree:context");
+
+/**
+ * A symbol for subscribing to events.
+ * @alpha
+ */
+export const on: unique symbol = Symbol("editable-tree:on");
 
 /**
  * A tree of an unknown type.
@@ -211,7 +235,35 @@ export interface UntypedTreeContext extends ISubscribable<ForestEvents> {
 	clear(): void;
 }
 
-{
-	type _check1 = requireAssignableTo<EditableTree, UntypedTree>;
-	type _check2 = requireAssignableTo<EditableField, UntypedField>;
+/**
+ * A collection of events that can be raised by an {@link EditableTree}.
+ * These events are triggered while the internal data structures are being updated.
+ * Thus these events must not trigger reading of the anchorSet or forest.
+ *
+ * TODO:
+ * - Design how events should be ordered.
+ * - Include sub-deltas in events.
+ * - Add more events.
+ * - Have some events (or a way to defer events) until the tree can be read.
+ *
+ * @alpha
+ */
+export interface EditableTreeEvents {
+	/**
+	 * Raised when a specific EditableTree node is changing.
+	 * This includes its fields.
+	 * @param upPath - the path corresponding to the location of the node being changed, upward.
+	 * @param value - the new value stored in the node.
+	 */
+	changing(upPath: UpPath): void;
+
+	/**
+	 * Raised when something in the tree is changing, including this node and its descendants.
+	 * The event can optionally return a {@link PathVisitor} to traverse the subtree
+	 * This event is called on every parent (transitively) when a change is occurring.
+	 * Includes changes to this node itself.
+	 * @param upPath - the path corresponding to the location of the node being changed, upward.
+	 * @returns a visitor to traverse the subtree or `void`.
+	 */
+	subtreeChanging(upPath: UpPath): PathVisitor | void;
 }
